# packages/tesslate-marketplace/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Tesslate Federated Marketplace `/v1`
  version: v1
  description: |
    Federated marketplace wire protocol for OpenSail / Tesslate Studio.

    Every response includes `X-Tesslate-Hub-Id` and
    `X-Tesslate-Hub-Api-Version: v1` headers. Endpoints whose capability
    isn't advertised in `/v1/manifest.capabilities[]` return a typed
    `501 unsupported_capability` JSON envelope.
  contact:
    name: Tesslate
    email: marketplace@tesslate.com
servers:
  - url: http://localhost:8800
    description: Local dev
  - url: https://marketplace.tesslate.com
    description: Tesslate Official
...
paths:
...
  /v1/items/{kind}/{slug}/pricing:
    get:
      tags: [pricing]
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Pricing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingDetail"
  /v1/items/{kind}/{slug}/checkout:
    post:
      tags: [pricing]
      security: []
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    Kind:
      in: path
      name: kind
      required: true
      schema:
        type: string
        enum: [agent, skill, mcp_server, base, app, theme, workflow_template]
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string }
...
  schemas:
    Pricing:
      type: object
      properties:
        pricing_type:
          type: string
          enum: [free, paid, subscription]
        price_cents: { type: integer }
        currency: { type: string }
        stripe_price_id: { type: string, nullable: true }
        interval: { type: string, nullable: true }
        extras: { type: object, additionalProperties: true, nullable: true }
...
    PricingDetail:
      type: object
      properties:
        pricing: { $ref: "#/components/schemas/Pricing" }
        listings:
          type: array
          items: { type: object, additionalProperties: true }
    CheckoutRequest:
      type: object
      properties:
        customer_email: { type: string, nullable: true }
        success_url: { type: string, nullable: true }
        cancel_url: { type: string, nullable: true }
        metadata:
          type: object
          additionalProperties: { type: string }
    CheckoutResponse:
      type: object
      properties:
        checkout_url: { type: string }
        session_id: { type: string }
        mode:
          type: string
          enum: [live, dev_simulator]
        expires_at: { type: string, format: date-time, nullable: true }
...